move EvictingSender into the streamer crate for reuse
