+++ streamer/src/lib.rs
@@ -1,4 +1,5 @@
 #![allow(clippy::arithmetic_side_effects)]
+pub mod evicting_sender;
 pub mod msghdr;
 pub mod nonblocking;
 pub mod packet;
